+++ cmd/bnsd/app/app.go
@@ -12,6 +12,10 @@ import (
 
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/app"
+	"github.com/iov-one/weave/cmd/bnsd/x/nft/blockchain"
+	"github.com/iov-one/weave/cmd/bnsd/x/nft/bootstrap_node"
+	"github.com/iov-one/weave/cmd/bnsd/x/nft/ticker"
+	"github.com/iov-one/weave/cmd/bnsd/x/nft/username"
 	"github.com/iov-one/weave/orm"
 	"github.com/iov-one/weave/store/iavl"
 	"github.com/iov-one/weave/x"
@@ -22,10 +26,6 @@ import (
 	"github.com/iov-one/weave/x/multisig"
 	"github.com/iov-one/weave/x/nft"
 	"github.com/iov-one/weave/x/nft/base"
-	"github.com/iov-one/weave/x/nft/blockchain"
-	"github.com/iov-one/weave/x/nft/bootstrap_node"
-	"github.com/iov-one/weave/x/nft/ticker"
-	"github.com/iov-one/weave/x/nft/username"
 	"github.com/iov-one/weave/x/sigs"
 	"github.com/iov-one/weave/x/utils"
 	"github.com/iov-one/weave/x/validators"
