+++ cmd/bnsd/x/nft/bootstrap_node/codec.proto
