+++ cmd/bnsd/x/nft/bootstrap_node/handler.go
@@ -2,11 +2,11 @@ package bootstrap_node
 
 import (
 	"github.com/iov-one/weave"
+	"github.com/iov-one/weave/cmd/bnsd/x/nft/blockchain"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/orm"
 	"github.com/iov-one/weave/x"
 	"github.com/iov-one/weave/x/nft"
-	"github.com/iov-one/weave/x/nft/blockchain"
 )
 
 const (
