+++ cmd/bnsd/x/nft/bootstrap_node/msg.go
