+++ cmd/bnsd/x/nft/ticker/codec.proto
