+++ cmd/bnsd/x/nft/ticker/init.go
