+++ cmd/bnsd/x/nft/ticker/init_test.go
