+++ cmd/bnsd/x/nft/ticker/msg.go
