+++ cmd/bnsd/x/nft/username/handler.go
@@ -2,11 +2,11 @@ package username
 
 import (
 	"github.com/iov-one/weave"
+	"github.com/iov-one/weave/cmd/bnsd/x/nft/blockchain"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/orm"
 	"github.com/iov-one/weave/x"
 	"github.com/iov-one/weave/x/nft"
-	"github.com/iov-one/weave/x/nft/blockchain"
 	"github.com/tendermint/tendermint/libs/common"
 )
