+++ cmd/bnsd/x/nft/username/model.go
@@ -4,10 +4,10 @@ import (
 	"bytes"
 
 	"github.com/iov-one/weave"
+	"github.com/iov-one/weave/cmd/bnsd/x/nft/blockchain"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/orm"
 	"github.com/iov-one/weave/x/nft"
-	"github.com/iov-one/weave/x/nft/blockchain"
 )
 
 type Token interface {
